requirements.txt
python-bitcoinrpc==1.0
python-dotenv==1.0.1
flet>=0.25.2